java_jni/refactoring: move code related to enum into separate module
