+++ macroslib/src/java_jni/java_code.rs
@@ -1,11 +1,12 @@
-use std::{fmt, io::Write, path::Path};
+use std::{io::Write, path::Path};
 
 use bitflags::bitflags;
 use rustc_hash::FxHashSet;
 use smol_str::SmolStr;
 
 use super::{
-    method_name, JavaConverter, JavaForeignTypeInfo, JniForeignMethodSignature, NullAnnotation,
+    map_write_err, method_name, JavaConverter, JavaForeignTypeInfo, JniForeignMethodSignature,
+    NullAnnotation,
 };
 use crate::{
     file_cache::FileWriteCache,
@@ -14,7 +15,7 @@ use crate::{
         ast::if_result_return_ok_err_types, TypeMap, FROM_VAR_TEMPLATE, TO_VAR_TEMPLATE,
         TO_VAR_TYPE_TEMPLATE,
     },
-    types::{ForeignEnumInfo, ForeignInterface, ForeignerClassInfo, MethodAccess, MethodVariant},
+    types::{ForeignInterface, ForeignerClassInfo, MethodAccess, MethodVariant},
     WRITE_TO_MEM_FAILED_MSG,
 };
 
@@ -26,62 +27,6 @@ bitflags! {
         const INTERNAL = 4;
         const COMMA_BEFORE = 8;
     }
-}
-
-pub(in crate::java_jni) fn generate_java_code_for_enum(
-    output_dir: &Path,
-    package_name: &str,
-    enum_info: &ForeignEnumInfo,
-) -> Result<(), String> {
-    let path = output_dir.join(format!("{}.java", enum_info.name));
-    let mut file = FileWriteCache::new(&path);
-    let enum_doc_comments = doc_comments_to_java_comments(&enum_info.doc_comments, true);
-    write!(
-        file,
-        r#"// Automaticaly generated by rust_swig
-package {package_name};
-
-{doc_comments}
-public enum {enum_name} {{
-"#,
-        package_name = package_name,
-        enum_name = enum_info.name,
-        doc_comments = enum_doc_comments,
-    )
-    .map_err(&map_write_err)?;
-
-    for (i, item) in enum_info.items.iter().enumerate() {
-        writeln!(
-            file,
-            "{doc_comments}{item_name}({index}){separator}",
-            item_name = item.name,
-            index = i,
-            doc_comments = doc_comments_to_java_comments(&item.doc_comments, false),
-            separator = if i == enum_info.items.len() - 1 {
-                ';'
-            } else {
-                ','
-            },
-        )
-        .map_err(&map_write_err)?;
-    }
-
-    write!(
-        file,
-        r#"
-    private int value;
-    {enum_name}(int value) {{
-        this.value = value;
-    }}
-    public final int getValue() {{ return value; }}
-}}
-"#,
-        enum_name = enum_info.name
-    )
-    .map_err(&map_write_err)?;
-
-    file.update_file_if_necessary().map_err(&map_write_err)?;
-    Ok(())
 }
 
 pub(in crate::java_jni) fn generate_java_code_for_interface(
@@ -628,7 +573,10 @@ fn convert_code_for_method<'a, NI: Iterator<Item = &'a str>>(
     (conv_code, args_for_call_internal)
 }
 
-fn doc_comments_to_java_comments(doc_comments: &[String], class_comments: bool) -> String {
+pub(in crate::java_jni) fn doc_comments_to_java_comments(
+    doc_comments: &[String],
+    class_comments: bool,
+) -> String {
     use std::fmt::Write;
     let mut comments = String::new();
     for (i, comment) in doc_comments.iter().enumerate() {
@@ -697,8 +645,4 @@ fn get_null_annotation_imports(
     }
 
     String::new()
-}
-
-fn map_write_err<Err: fmt::Display>(err: Err) -> String {
-    format!("write failed: {}", err)
 }
